"Naked" blocks must begin with "<" and end with ">"
I had the "hey, here's a block here already!" detection set to only
check for a tag at the end of the block, but realized that that could
mess things up if a block ends on, say, an emphasized bit of text.

This change checks both that the block begins with a tag and ends with a
tag, which can still give false positives (and could engender some false
negatives as well), but is, in my opinion, better overall.